requirements.txt
tweepy>=4.14.0
praw>=7.7.0
feedparser>=6.0.10
anthropic>=0.25.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
click>=8.1.0
loguru>=0.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
